Rename MORR.Core.Data.Transcoding.WinAPI to MORR.Core.Data.Transcoding.MPEG and remove unused Decoder
