+++ application/Core/MORR/Data/Transcoding/MPEG/MPEGEncoderConfiguration.cs
@@ -4,7 +4,7 @@ using MORR.Core.Configuration;
 using MORR.Shared.Configuration;
 using MORR.Shared.Utility;
 
-namespace MORR.Core.Data.Transcoding.WinAPI
+namespace MORR.Core.Data.Transcoding.MPEG
 {
     public class MPEGEncoderConfiguration : IConfiguration
     {
