--- application/Core/MORR/Data/Transcoding/WinAPI/Decoder.cs
@@ -1,15 +1,0 @@
-﻿﻿using MORR.Shared.Utility;
-
-namespace MORR.Core.Data.Transcoding.WinAPI
-{
-    /// <summary>
-    ///     Decodes an MPEG-Stream using the Windows API
-    /// </summary>
-    public class Decoder : IDecoder
-    {
-        public void Decode(FilePath path)
-        {
-            throw new System.NotImplementedException();
-        }
-    }
-}
